requirements.txt
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
greenlet==2.0.2
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
mysql-connector-python==8.0.33
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
protobuf==3.20.3
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.16
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
wcwidth==0.2.6
xlrd==2.0.1
